chore!: DBTP-905 Remove bootstrap commands and requirement for bootstrap.yml (#402)
