+++ dbt_platform_helper/utils/application.py
@@ -10,9 +10,7 @@ from yaml.parser import ParserError
 
 from dbt_platform_helper.utils.aws import get_aws_session_or_abort
 from dbt_platform_helper.utils.aws import get_profile_name_from_account_id
-from dbt_platform_helper.utils.files import load_and_validate_config
 from dbt_platform_helper.utils.messages import abort_with_error
-from dbt_platform_helper.utils.validation import BOOTSTRAP_SCHEMA
 
 
 class Environment:
@@ -111,21 +109,13 @@ def load_application(app: str = None, default_session: Session = None) -> Applic
 
 def get_application_name():
     app_name = None
-
     try:
-        app_config = load_and_validate_config("bootstrap.yml", BOOTSTRAP_SCHEMA)
-        app_name = app_config["app"]
-    except (FileNotFoundError, ParserError):
-        pass
-
-    try:
-        if app_name is None:
-            app_config = yaml.safe_load(Path("copilot/.workspace").read_text())
-            app_name = app_config["application"]
+        app_config = yaml.safe_load(Path("copilot/.workspace").read_text())
+        app_name = app_config["application"]
     except (FileNotFoundError, ParserError):
         pass
 
     if app_name is None:
-        abort_with_error("No valid bootstrap.yml or copilot/.workspace file found")
+        abort_with_error("Cannot get application name. No copilot/.workspace file found")
 
     return app_name
